move add_prefix, groupby cov corr
